pyproject.toml
[project]
name = "search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.9.3",
    "fastmcp>=2.13.3",
    "langchain>=1.1.3",
    "langchain-core>=1.1.3",
    "langchain-openai>=1.1.1",
    "langgraph>=1.0.4",
    "openai>=2.9.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "trafilatura>=2.0.0",
]
